+++ libs/ferriskey-organization/Cargo.toml
@@ -15,6 +15,8 @@ mockall = { version = "0.14.0", optional = true }
 
 [dev-dependencies]
 mockall = "0.14.0"
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros"] }
+ferriskey-domain = { path = "../ferriskey-domain", features = ["mock"] }
 
 [features]
 mock = ["mockall"]
