+++ libs/ferriskey-organization/src/lib.rs
@@ -4,4 +4,6 @@ pub mod ports;
 pub use entities::*;
 pub use ports::*;
 
+pub mod group_services;
 pub mod policies;
+pub mod services;
